+++ tmp/request.hpp
@@ -32,21 +32,18 @@ private:
     std::string parseMethod(const std::string& methodStr);
 };
 
-void parse(const std::string& request)
+void HttpRequest::parse(const std::string& request)
 {
 	std::istringstream iss(request);
 	std::string line;
 
-	// 1. 요청 라인 파싱
 	std::getline(iss, line);
 	parseRequestLine(line);
 
-	// 2. 요청 헤더 파싱
 	while (std::getline(iss, line) && !line.empty()) {
 		parseHeader(line);
 	}
 
-	// 3. 요청 본문 파싱 (선택적)
 	if (headers.find("Content-Length") != headers.end()) {
 		int contentLength = atoi(headers["Content-Length"].c_str());
 		std::string body(contentLength, '\0');
@@ -55,17 +52,17 @@ void parse(const std::string& request)
 	}
 }
 
-const std::string& getMethod() const { return method; }
+const std::string& HttpRequest::getMethod() const { return method; }
 
-const std::string& getUri() const { return uri; }
+const std::string& HttpRequest::getUri() const { return uri; }
 
-const std::string& getVersion() const { return version; }
+const std::string& HttpRequest::getVersion() const { return version; }
 
-const std::map<std::string, std::string>& getHeaders() const { return headers; }
+const std::map<std::string, std::string>& HttpRequest::getHeaders() const { return headers; }
 
-const std::string& getBody() const { return body; }
+const std::string& HttpRequest::getBody() const { return body; }
 
-void parseRequestLine(const std::string& line)
+void HttpRequest::parseRequestLine(const std::string& line)
 {
     std::istringstream iss(line);
 	std::string token;
@@ -78,7 +75,7 @@ void parseRequestLine(const std::string& line)
 	iss >> version;
 }
 
-std::string parseMethod(const std::string& methodStr)
+std::string HttpRequest::parseMethod(const std::string& methodStr)
 {
 	if (methodStr == "GET")
 		return "GET";
@@ -90,7 +87,7 @@ std::string parseMethod(const std::string& methodStr)
 		return "OTHER";
 }
 
-void parseHeader(const std::string& line)
+void HttpRequest::parseHeader(const std::string& line)
 {
 	std::string::size_type pos = line.find(":");
 	if (pos != std::string::npos) {
